requirements.txt
ruamel.yaml
GitPython
requests
